+++ config.env
@@ -1,0 +1,9 @@
+NODE_ENV=development
+MONGO_URI=mongodb://localhost:27017/salechampnl      
+JWT_SECRET=        
+JWT_EXPIRE=       
+JWT_COOKIE_EXPIRE=      
+SMTP_HOST=
+SMTP_EMAIL=        
+SMTP_PASSWORD=
+PORT=5000
