+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "project-startup",
   "version": "1.0.0",
-  "description": "Project startup for node.js",
+  "description": "Saleschamp",
   "main": "server.js",
   "scripts": {
-    "start:dev": "nodemon server.js",
-    "start:prod": "NODE_ENV=production nodemon server.js",
+    "start": "nodemon  ./src/server.js",
+    "start:prod": "NODE_ENV=production nodemon ./src/server.js",
+    "start:dev": "NODE_ENV=development nodemon ./src/server.js",
     "debug": "ndb server.js"
   },
   "keywords": [
@@ -20,19 +21,11 @@
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
-    "ejs": "^3.1.6",
     "express": "^4.17.1",
-    "express-mongo-sanitize": "^2.0.0",
-    "express-rate-limit": "^5.1.3",
-    "helmet": "^4.1.1",
-    "hpp": "^0.2.3",
     "i18n-iso-countries": "^7.3.0",
     "jest": "^27.5.1",
-    "moment": "^2.29.1",
     "mongoose": "^5.13.5",
     "morgan": "^1.10.0",
-    "supertest": "^6.2.2",
-    "xss": "^1.0.8",
-    "xss-clean": "^0.1.1"
+    "supertest": "^6.2.2"
   }
 }
