+++ src/config/db.js
@@ -1,18 +1,13 @@
 const mongoose = require("mongoose");
-const keys = require("./keys");
 
 const connectDB = async () => {
-  const conn = await mongoose.connect(keys.MONGO_URI, {
+  const conn = await mongoose.connect(process.env.MONGO_URI, {
     useNewUrlParser: true,
     useCreateIndex: true,
     useFindAndModify: false,
     useUnifiedTopology: true
   });
-
-  console.log(
-    `MongoDB Connected: ${conn.connection.host}.saleschamp.nl-dev`.yellow
-      .underline.bold
-  );
+  console.log(`MongoDB Connected ---SalesChamp`);
 };
 
 module.exports = connectDB;
