+++ src/middleware/error.js
@@ -1,4 +1,4 @@
-const ErrorResponse = require('../utils/errorResponse');
+const ErrorResponse = require("../utils/errorResponse");
 
 const errorHandler = (err, req, res, next) => {
   let error = { ...err };
@@ -9,26 +9,26 @@ const errorHandler = (err, req, res, next) => {
   console.log(err);
 
   // Mongoose bad ObjectId
-  if (err.name === 'CastError') {
+  if (err.name === "CastError") {
     const message = `Resource not found`;
     error = new ErrorResponse(message, 404);
   }
 
   // Mongoose duplicate key
   if (err.code === 11000) {
-    const message = 'Duplicate field value entered';
+    const message = "Duplicate field value entered";
     error = new ErrorResponse(message, 400);
   }
 
   // Mongoose validation error
-  if (err.name === 'ValidationError') {
+  if (err.name === "ValidationError") {
     const message = Object.values(err.errors).map(val => val.message);
     error = new ErrorResponse(message, 400);
   }
 
   res.status(error.statusCode || 500).json({
     success: false,
-    error: error.message || 'Server Error'
+    error: error.message || "Server Error"
   });
 };
