+++ src/models/Address.js
@@ -66,9 +66,4 @@ const AddressSchema = new mongoose.Schema({
   }
 });
 
-
-AddressSchema.pre('save' ,function(){
-  
-})
-
 module.exports = mongoose.model("Address", AddressSchema);
